+++ sample/sample12-custom-naming-strategy/naming-strategy/CustomNamingStrategy.ts
@@ -1,9 +1,7 @@
 import {NamingStrategyInterface} from "../../../src/naming-strategy/NamingStrategyInterface";
-import {NamingStrategy} from "../../../src/decorator/NamingStrategy";
 import {DefaultNamingStrategy} from "../../../src/naming-strategy/DefaultNamingStrategy";
 import {snakeCase} from "../../../src/util/StringUtils";
 
-@NamingStrategy("custom_strategy")
 export class CustomNamingStrategy extends DefaultNamingStrategy implements NamingStrategyInterface {
 
     tableName(targetName: string, userSpecifiedName: string): string {