+++ src/driver/DriverType.ts
@@ -1,0 +1,4 @@
+/**
+ * Driver type.
+ */
+export type DriverType = "mysql"|"postgres"|"mariadb"|"sqlite"|"oracle"|"mssql"|"websql"|"mongodb";