+++ src/driver/mongodb/MongoDriver.ts
@@ -24,16 +24,6 @@ export class MongoDriver implements Driver {
     // -------------------------------------------------------------------------
 
     /**
-     * Naming strategy used in the connection where this driver is used.
-     */
-    namingStrategy: NamingStrategyInterface;
-
-    /**
-     * Used to wrap lazy relations to be able to perform lazy loadings.
-     */
-    lazyRelationsWrapper: LazyRelationsWrapper;
-
-    /**
      * Mongodb does not require to dynamically create query runner each time,
      * because it does not have a regular pool.
      */