+++ src/metadata-builder/JunctionEntityMetadataBuilder.ts
@@ -30,7 +30,7 @@ export class JunctionEntityMetadataBuilder {
         const referencedColumns = this.collectReferencedColumns(relation, joinTable);
         const inverseReferencedColumns = this.collectInverseReferencedColumns(relation, joinTable);
 
-        const joinTableName = joinTable.name || this.connection.driver.namingStrategy.joinTableName(
+        const joinTableName = joinTable.name || this.connection.namingStrategy.joinTableName(
             relation.entityMetadata.tableNameWithoutPrefix,
             relation.inverseEntityMetadata.tableNameWithoutPrefix,
             relation.propertyPath,
@@ -52,7 +52,7 @@ export class JunctionEntityMetadataBuilder {
                 return (!joinColumnArgs.referencedColumnName || joinColumnArgs.referencedColumnName === referencedColumn.propertyName) &&
                     !!joinColumnArgs.name;
             }) : undefined;
-            const columnName = joinColumn && joinColumn.name ? joinColumn.name : this.connection.driver.namingStrategy.joinTableColumnName(relation.entityMetadata.tableNameWithoutPrefix, referencedColumn.propertyName, referencedColumn.databaseName);
+            const columnName = joinColumn && joinColumn.name ? joinColumn.name : this.connection.namingStrategy.joinTableColumnName(relation.entityMetadata.tableNameWithoutPrefix, referencedColumn.propertyName, referencedColumn.databaseName);
 
             return new ColumnMetadata({
                 entityMetadata: entityMetadata,
@@ -78,7 +78,7 @@ export class JunctionEntityMetadataBuilder {
                 return (!joinColumnArgs.referencedColumnName || joinColumnArgs.referencedColumnName === inverseReferencedColumn.propertyName) &&
                     !!joinColumnArgs.name;
             }) : undefined;
-            const columnName = joinColumn && joinColumn.name ? joinColumn.name : this.connection.driver.namingStrategy.joinTableColumnName(relation.inverseEntityMetadata.tableNameWithoutPrefix, inverseReferencedColumn.propertyName, inverseReferencedColumn.databaseName);
+            const columnName = joinColumn && joinColumn.name ? joinColumn.name : this.connection.namingStrategy.joinTableColumnName(relation.inverseEntityMetadata.tableNameWithoutPrefix, inverseReferencedColumn.propertyName, inverseReferencedColumn.databaseName);
 
             return new ColumnMetadata({
                 entityMetadata: entityMetadata,