+++ src/naming-strategy/DefaultNamingStrategy.ts
@@ -13,7 +13,7 @@ export class DefaultNamingStrategy implements NamingStrategyInterface {
      * @param targetName Name of the target entity that can be used to generate a table name.
      * @param userSpecifiedName For example if user specified a table name in a decorator, e.g. @Entity("name")
      */
-    tableName(targetName: string, userSpecifiedName: string): string {
+    tableName(targetName: string, userSpecifiedName: string|undefined): string {
         return userSpecifiedName ? userSpecifiedName : snakeCase(targetName);
     }